+++ next.config.js
@@ -108,5 +108,5 @@ const moduleExports = {
 };
 
 module.exports = withSentryConfig(moduleExports, {
-  dryRun: process.env.VERCEL_ENV !== "production",
+  dryRun: true,
 });