+++ src/config/chain/types.ts
@@ -53,7 +53,7 @@ type GovConfig =
 type NftConfig = { enabled: boolean };
 
 export interface ChainConfig {
-  tier: "lite" | "full";
+  tier: "lite" | "mesa" | "sequencer" | "full";
   chain: string;
   registryChainName: string;
   prettyName: string;