+++ src/lib/services/block/sequencer.ts
@@ -1,0 +1,30 @@
+import axios from "axios";
+
+import {
+  zBlockDataResponseSequencer,
+  zBlocksResponseSequencer,
+} from "../types";
+import type { Option } from "lib/types";
+import { parseWithError } from "lib/utils";
+
+export const getBlocksSequencer = async (
+  endpoint: string,
+  paginationKey: Option<string>
+) =>
+  axios
+    .get(`${endpoint}/indexer/block/v1/blocks`, {
+      params: {
+        "pagination.offset": 0,
+        "pagination.limit": 10,
+        "pagination.reverse": true,
+        "pagination.key": paginationKey,
+      },
+    })
+    .then(({ data }) => parseWithError(zBlocksResponseSequencer, data));
+
+export const getBlockDataSequencer = async (endpoint: string, height: number) =>
+  axios
+    .get(`${endpoint}/indexer/block/v1/blocks/${height}`)
+    .then(({ data }) =>
+      parseWithError(zBlockDataResponseSequencer, data.block)
+    );